Renamed the "experiment" subdirectory to "experimental"
This was the original name, but it seemed too long compared to the
other directories, so it was shortened down to "experiment". But the
name "experimental" still seems more appropriate, so it is being
changed back to that term. Sorry for any confusion.
