+++ parts/experimental/C1BitCPU.java
@@ -1,4 +1,4 @@
-package parts.experiment;
+package parts.experimental;
 import parts.*;
 
 /**
